+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.0.1] - 2025-05-13
+
 ### Fixed
 
 - Fix limits default value in mojos
@@ -18,5 +20,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Initial release.
 
-[Unreleased]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/nbbrd/nbbrd-maven-tools/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/nbbrd/nbbrd-maven-tools/compare/develop...v1.0.0